% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderInterface.R
\name{SMRMakeServerFunction}
\alias{SMRMakeServerFunction}
\title{Sparse matrix recommender interface server function}
\usage{
SMRMakeServerFunction(itemSMR, itemData, itemDataColNames = NULL,
  itemDataIDColName = NULL, searchColName = NULL,
  itemListIDsSplitPattern = "\\\\W")
}
\arguments{
\item{itemSMR}{A time series recommender.}

\item{itemData}{A data frame with rows corresponding to items (to be recommended.)}

\item{itemDataColNames}{Which column names of \code{itemData} should be used in
the table displays of search and recommendation results.}

\item{itemDataIDColName}{Which column of \code{itemData} is with item ID's;
those ID's are also row names of \code{itemSMR$M}.}

\item{searchColName}{Which column should be used to search \code{itemData}.}

\item{itemListIDsSplitPattern}{A split pattern for the separator of the items list
and ratings list.}
}
\value{
Shiny server function.
}
\description{
Creates the Shiny server function for a sparse matrix recommender interface.
}
\details{
The default value of \code{itemListIDsSplitPattern} is '\\W', but "," should be
used if the row ID's have white spaces in them.
}
\seealso{
Other SMR interface functions: \code{\link{SMRCreateSearchInterface}},
  \code{\link{SMRMakeUI}},
  \code{\link{SMRMonMakeServerFunction}}
}
\concept{SMR interface functions}
