% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrixRecommenderInterface.R
\name{SMRMakeUI}
\alias{SMRMakeUI}
\title{Sparse matrix recommender interface UI}
\usage{
SMRMakeUI(itemSMR)
}
\arguments{
\item{itemSMR}{A sparse matrix recommender.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for a sparse matrix recommender interface.
}
\seealso{
Other SMR interface functions: \code{\link{SMRCreateSearchInterface}},
  \code{\link{SMRMakeServerFunction}},
  \code{\link{SMRMonMakeServerFunction}}
}
\concept{SMR interface functions}
