% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMonInterface.R
\name{SMRMonCreateInteractiveInterface}
\alias{SMRMonCreateInteractiveInterface}
\title{Creation of a SMRMon interface}
\usage{
SMRMonCreateInteractiveInterface(smrObj, itemData,
  itemDataColNames = NULL, itemDataIDColName = NULL,
  searchColName = NULL, itemListIDsSplitPattern = "\\\\W",
  dashboardTheme = NULL)
}
\arguments{
\item{itemData}{A data frame with rows corresponding to items (to be recommended.)}

\item{itemDataColNames}{Which column names of \code{itemData} should be used in
the search and recommendation results.}

\item{itemDataIDColName}{Which column of \code{itemData} is with item ID's;
those ID's are also row names of \code{smrObj$M}.}

\item{itemListIDsSplitPattern}{A split pattern for the separator of the items list
and ratings list.}

\item{dashboardTheme}{A string specifying the dashbaard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}

\item{smarObj}{A SMRMon object.}
}
\value{
Shiny app
}
\description{
Creation of a SMRMon interface
}
\details{
The default value of \code{itemListIDsSplitPattern} is '\\W', but "," should be
used if the row ID's have white spaces in them.
}
\seealso{
Other SMRMon interface functions: \code{\link{SMRMonMakeUI}}
}
\concept{SMRMon interface functions}
