% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMonSearchInterface.R
\name{SMRMonMakeSearchServerFunction}
\alias{SMRMonMakeSearchServerFunction}
\title{SMRMon search interface server function}
\usage{
SMRMonMakeSearchServerFunction(
  smrObj,
  itemData,
  itemDataColNames = NULL,
  itemDataIDColName = NULL,
  itemListIDsSplitPattern = "\\\\W"
)
}
\arguments{
\item{smrObj}{A time series recommender.}

\item{itemData}{A data frame with rows corresponding to items (to be recommended.)}

\item{itemDataColNames}{Which column names of \code{itemData} should be used in
the table displays of search and recommendation results.}

\item{itemDataIDColName}{Which column of \code{itemData} is with item ID's;
those ID's are also row names of \code{smrObj$M}.}

\item{itemListIDsSplitPattern}{A split pattern for the separator of the items list
and ratings list.}
}
\value{
Shiny server function.
}
\description{
Creates the Shiny server function for a sparse matrix recommender interface.
}
\details{
The default value of \code{itemListIDsSplitPattern} is '\\W', but "," should be
used if the row ID's have white spaces in them.
}
\seealso{
Other SMRMon interface functions: 
\code{\link{SMRMonCreateInteractiveInterface}()},
\code{\link{SMRMonCreateSearchInteractiveInterface}()},
\code{\link{SMRMonMakeSearchUI}()},
\code{\link{SMRMonMakeUI}()}
}
\concept{SMRMon interface functions}
