% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMonSearchInterface.R
\name{SMRMonMakeSearchUI}
\alias{SMRMonMakeSearchUI}
\title{SMRMon interface UI}
\usage{
SMRMonMakeSearchUI(smrObj, dashboardTheme = NULL)
}
\arguments{
\item{smrObj}{A SMRMon object.}

\item{dashboardTheme}{A string specifying the dashbaard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for a sparse matrix recommender SMRMon interface.
}
\seealso{
Other SMRMon interface functions: 
\code{\link{SMRMonCreateInteractiveInterface}()},
\code{\link{SMRMonCreateSearchInteractiveInterface}()},
\code{\link{SMRMonMakeSearchServerFunction}()},
\code{\link{SMRMonMakeUI}()}
}
\concept{SMRMon interface functions}
