% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRMonInterface.R
\name{SMRMonMakeUI}
\alias{SMRMonMakeUI}
\title{SMRMon interface UI}
\usage{
SMRMonMakeUI(smrObj, dashboardTheme = NULL)
}
\arguments{
\item{smrObj}{A SMRMon object.}

\item{dashboardTheme}{A string specifying the dashbaard theme.
See \code{\link{dashboardthemes::shinyDashboardThemes}}.
Ignored if the package \code{dashboardthemes} is not installed.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for a sparse matrix recommender SMRMon interface.
}
\seealso{
Other SMRMon interface functions: \code{\link{SMRMonCreateInteractiveInterface}}
}
\concept{SMRMon interface functions}
