% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRCreateSearchInterface}
\alias{TSCorrSMRCreateSearchInterface}
\title{Creation of time series search interface}
\usage{
TSCorrSMRCreateSearchInterface(tsSMR, tsSearchVectors = NULL, roundDigits = 6)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}

\item{roundDigits}{Number of decimal places for \code{\link{round}}.
(Used for making the \code{ggplot} panel names.)}
}
\value{
Shiny app object.
}
\description{
Creates Shiny interface for a given time series recommender and
optional search time series shapes.
}
\seealso{
Other Time series search interface functions: 
\code{\link{TSCorrSMRMakeServerFunction}()},
\code{\link{TSCorrSMRMakeUI}()}
}
\concept{Time series search interface functions}
