% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRMakeUI}
\alias{TSCorrSMRMakeUI}
\title{Time series search interface UI}
\usage{
TSCorrSMRMakeUI(tsSMR, tsSearchVectors)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}
}
\value{
Shiny UI object.
}
\description{
Creates the Shiny UI function for a time series search interface.
}
\seealso{
Other Time series search interface functions: \code{\link{TSCorrSMRCreateSearchInterface}},
  \code{\link{TSCorrSMRMakeServerFunction}}
}
\concept{Time series search interface functions}
