% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRCreateSearchInterface}
\alias{TSCorrSMRCreateSearchInterface}
\title{Creation of time series search interface}
\usage{
TSCorrSMRCreateSearchInterface(tsSMR, tsSearchVectors = NULL)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}
}
\value{
Shiny app object.
}
\description{
Creates Shiny interface for a given time series recommender and
optional search time series shapes.
}
\seealso{
Other Time series search interface functions: \code{\link{SMRMakeUI}},
  \code{\link{TSCorrSMRMakeServerFunction}},
  \code{\link{TSCorrSMRMakeUI}}
}
\concept{Time series search interface functions}
