% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesSearchInterface.R
\name{TSCorrSMRMakeServerFunction}
\alias{TSCorrSMRMakeServerFunction}
\title{Time series search interface server function}
\usage{
TSCorrSMRMakeServerFunction(tsSMR, tsSearchVectors, roundDigits = 6)
}
\arguments{
\item{tsSMR}{A time series recommender.}

\item{tsSearchVectors}{A list of time series search vectors.}

\item{roundDigits}{Number of decimal places for \code{\link{round}}.
(Used for making the \code{ggplot} panel names.)}
}
\value{
Shiny server function.
}
\description{
Creates the Shiny server function for a time series search interface.
}
\seealso{
Other Time series search interface functions: 
\code{\link{TSCorrSMRCreateSearchInterface}()},
\code{\link{TSCorrSMRMakeUI}()}
}
\concept{Time series search interface functions}
